Remove now redundant getRootFromIndexLink helper
