+++ client/src/onload/loadConfig.js
@@ -1,13 +1,11 @@
-// loadConfig.ts
-import { getRootFromIndexLink } from "@/onload/getRootFromIndexLink";
+import { serverPath } from "@/utils/serverPath";
 
 let cachedConfig = null;
 
 export async function loadConfig() {
     if (!cachedConfig) {
         try {
-            const root = getRootFromIndexLink();
-            const response = await fetch(`${root}context`);
+            const response = await fetch(`${getAppRoot()}context`);
             if (!response.ok) {
                 throw new Error(`Failed to fetch /context (${response.status})`);
             }
@@ -20,6 +18,18 @@ export async function loadConfig() {
     return cachedConfig;
 }
 
-export function getAppRoot() {
-    return getRootFromIndexLink();
+/**
+ * Finds <link rel="index"> in head element and pulls root url fragment from
+ * there.
+ *
+ * @param {string} [defaultRoot="/"]
+ * @returns {string}
+ */
+export function getAppRoot(defaultRoot = "/") {
+    if (typeof document === "undefined") {
+        return defaultRoot;
+    }
+    const links = document.getElementsByTagName("link");
+    const indexLink = Array.from(links).find((link) => link.rel == "index");
+    return indexLink && indexLink.href ? serverPath(indexLink.href) : defaultRoot;
 }
